BufferedOutputStream: align CoreStream position
After reopening an BufferedOutputStream the underlying CoreStream
might be at an arbitrary position since the stream is purged on close.
To avoid unnecessary operations on the underlying stream,
e.g. because of slices crossing the block boundary we adjust
the length of the first slice to align the position to slice size
such that after the first write all following writes on the underlying
stream will be aligned to slice size again. This also helps to avoid
unaligned writes in block storage tiers.

https://issues.apache.org/jira/browse/CRAIL-17

Close #11

Signed-off-by: Jonas Pfefferle <pepperjo@apache.org>